16-10-25
- New: PK, standard - No replacer module.
It will add only the PK LMG and not remove the vanilla equivalents of the PKM and PKP. The Armor Modkits addon patch will upgrade the PK to the vanilla PKM.
- New: G28 module: MAS patch.
This will detach the fixed scope off the G28, allowing you to attach whatever you want. Given the lack of proper rear sights, the one from the XM8 has been used as a patchwork.
Using Armor Modkits addon and a specialized modkit, you can still upgrade it to the original version with the fixed PM-II scope and attach the PN-23 "Nozzle" night sight. You can also use a (much cheaper) gunsmith modkit to revert back to the unfixed version, but you need to detach the PN-23 first.
This patch also includes tweaks to the weapon's mass. Unscoped and empty will have the 5.8 kgs as in real life, while with the PM-II scope it'll weigh 7.1 instead of 4. The PN23 will raise it to 7.750, but also it will add to the gun's value.
- Update: PK, tactical variants module.
Armor Modkits addon - PK - No replacer patch. Allows upgrading the vanilla PKM and PKP to the tactical variants. To be used when no PK, standard option is present, or along the new PK, standard - No replacer option.
- Update: DXML'd the UI additions of the G28 and Karabiner 98 scoping textures for better compatibility (some mods that include the UI elements for BaS scopes don't do it for one or both of these two).
- Fix: G28 module. Crashes due to incorrectly configured scoping texture.