28-12-25
Removed useless trade_preset files.
- Fix: Scopes and PUSSY/PERIOD patch. Some icons didn't show up. I reviewed them all, and should now appear correctly.
- Fix: AK-105 module: Loris still selling the vanilla one with the replacer option.
- Update: No BaS addons patch. Up to date with Uncle Yar's Workshop version. To note, modules that keep certain addons will ensure certain gun mods keep using them as intended by the author (as long as they're known).
ie: ATHI's SIG MCX Spear uses the SCAR grenade launcher. This patch changes it for the vanilla M203. The SCAR module includes the SCAR gl, and will reapply it to the MCX Spear.
Other known cases are the MK18 Mjolnir and Gungnir with the M700 .338LM Sound Suppressor.